Disable vimple version numbers when using neovim
Vimple doesn't support neovim, and the version strings are *very*
different.
This causes an E688 upon startup unless we disable it.
Disable it as I don't use the version information for anything.
Only disable it in neovims' "init.vim" because that's where the problem
is, and it's nice to keep things separate.