Make '<RightMouse> pop the tag stack
Instead of using the default g<RightMouse>, I start the mapping with "'"
because "g" seems more fitting to "grep" to me.
I make sure to use 'nmap' rather than 'nnoremap' because that makes use
of any remappings of <C-T> that may occur.
It may also cause problems if people map <LeftMouse> in some plugins,
but for now I'm just hoping that doesn't happen.