Update customisation of 'K' binding for man pages
Neovim doesn't act in the same way as vim does in this regards, so if
running neovim, don't use the extra customisation in man_settings.vim.

Instead, have the default action for 'K' be set by 'keywordprg' in the
main vimrc, and let other plugins change it if they want.