QFRemoveCurrent partially accounts for extra text
Before we didn't account for extra error information in the quickfix
list (assuming each line in the list we get from getqflist() is a valid
error).
We now account for the extra information under items, but we can't
handle when the "current" quickfix item is one of these information
lines.