Stop Occur leaving temporary qflists in history
This command creates a temporary quickfix list before filtering it into
the final version presented to the user.
Before this patch the temporary list would stay around in the quickfix
histotry, which pollutes the :colder :cnewer histotry.

This patch removes that problem, but makes the organisation of Quickfix
filtering functions slightly more ugly.

I may have a look at that sometime in the future.