Don't have '**' in 'path'
I've been using this as an option for a while, but I'm not sure whether
it's actually providing benefit.
There are already fuzzy finder plugins, and without this setting plugins
can use 'path' in the way it was designed.

This is more of a test, to see if I find it annoying in my daily life.