OccurSearch now checks for quickfix title existing
As I only tested this in neovim I didn't notice that this was a problem.
It turns out that in vim the returned dictionary from
`getqflist({'title': 1})` can be empty.
The time this has been observed to happen is just after opening the
editor, and before having any errors in the quickfix list.

To account for this, `OccurSearch()` now uses `get()` with a default
when getting that property.