Add commands and mappings for handy spellfixing
The idea here is to create a commit with the spellfix, and get back to
what you were doing with the minimal required thought.

The functionality currently leaves commits around for you to have to
clean up later, which is a downside, but the alternative is for you to
have to save the spell fix in a sensible format now, which requires
thought now.

I'm thinking avoiding a context switch is the best way to approach this.