Occuradd should not short-circuit on quickfix title
The short-circuiting is a cache mechanism, which avoids work if the
given work has already been done.
Occuradd will change the state of the quickfix list on multiple
invokations.
Hence the cache mechanism is changing the behaviour of the command.

There is a slight change that the cache mechanism introduces for normal
:Occur, where the number of lists in the quickfix stack will not change
due to the caching.
This is a very small change, and a positive, so I'm not changing it.