Some improvements to TQFSelect and friends
I've found the meaning of a few more tag kinds, and hence updated the
data, I've also noticed that `ctags` can produce fully described tag
kinds and accounted for that.

On top of that I removed some unneeded functionality in the main
TQFSelect command.
There were options for a ! and for calling without arguments that I
honestly don't know what they were intended for.

Finally I fixed a bug in TQFSetQuickfix() function that demonstrated I
didn't even test the function after writing it.