Attempt to reduce the complexity of shifted numbers
As it stands there is just slightly too much variation in the shifted
numbers keymap between filetypes (not to mention between some files
where I've turned this off and some where I haven't).

Working like this can trigger a lot of mistyped variables, which negates
the whole point of typing "in the flow" that avoiding having to press
Shift was supposed to give.

Here I adjust settings such that the `_/-` mapping is always on when the
keymap is applied and remove all times when I switch between `"/'`.
The quote switch was not that beneficial anyway, since I was typing
contractions in comments nearly as often as I was typing
string-literals in C-like languages.