Ensure we set `mapleader` before `syntax on` in vimrc
After a recent neovim change `syntax on` in vimrc triggers loading
ftplugins for the filetype specified on the command line.  The `vsh`
ftplugins create mappings that go on <leader> and <localleader>.
Hence these mappings go on the `mapleader` and `maplocalleader` that are
specified at the time of seeing `syntax on` in vimrc if either a python
or vsh file are loaded from the command line.

If these are not specified at the time that `syntax on` is seen in the
vimrc then vim uses the default of `\` for both.

There's not really anything that vsh can do for this.  The plugin wants
mappings on <Leader> and <localleader>.  When those are not defined they
should go on the default as per vim behaviour.  There's no way to tell
from the plugin whether `mapleader` was going to be defined later on, or
that the defaults were what the end-user wanted.

Hence adaptation for this behaviour must be made by an end users vimrc.
This adaptation is to ensure that `mapleader` and `maplocalleader` are
specified before `syntax on` if `syntax on` is in the users vimrc.

(For ordering behaviour see https://github.com/neovim/neovim/issues/19008)