Fix bug in RemoveCurrentQuickfixItem
When trying to identify the "current" quickfix item we were attempting
to match the column that we end up at after running `cc` with the column
in the quickfix entry.

Unfortunately we were identifying the column we end up in using
`virtcol`, which is not what the `col` entry in quickfix items
represents.

This meant that whenever these two values differed (e.g. if there is a
TAB character in the relevant line) we would not find the relevant
entry.

Here we fix that by using `col` instead of `virtcol` to match the values
stored in quickfix entries.