RemoveCurrentQuickfixItem account for QF items using pattern
There are two mechanisms by which a quickfix item can specify the line
to go to.  It can either directly specify the line number or it can
specify the pattern to search for.

When specifying the pattern to search for the line number is given as 0,
which doesn't match any line in the file.  Hence we need to use that
pattern when trying to find the "current" quickfix item.

This commit adds the ability to use the pattern rather than line number
when attempting to remove the "current" quickfix item.