Add LSP configuration
As yet very basic, but it seems to be giving something.
Will need to do something about starting the correct server for the type
of file that I'm in etc (especially if I end up with multiple projects
open at the same time), but for the moment this gives at least some
functionality that I didn't have before.