Treesitter enable highlighting for smaller files
Treesitter provides a function to check whether to enable syntax
highlighting or not.  This allows me to turn the syntax highlighting off
for large C/C++ files.  These files were seeing some slowdown during
typing when I had treesitter syntax highlighting set up.

Also define a custom foldtext function based on treesitter.  This
improves upon the default foldtext (of the first line) especially for
GNU style C code which has the return type of the function on a separate
line to the rest of the function declaration.