Adjust what gets folded rather than the foldtext
This gives a much more similar behaviour to `foldmethod=syntax` in that
the function body is what gets folded rather than the entire function
definition.

I *think* I like this more, getting a bit more of an overview of what's
going on.  Honestly not sure, going with this for the moment and will
see if adjusting makes sense later.