Do not call `syntax on` or `filetype on` in vimrc for neovim
These were specifically mentioned as should not be called in vimrc
upstream, and having them in my vimrc seems to make a few things not
work very well.

Removing them *seems* to make treesitter work on the first file open,
and also avoid a mappings difficulty I had with Vsh.