Install wilder.nvim
I played around with telescope-cmdline, but found a few reasons this
wasn't working for me.  Looks like `wilder` is doing exactly what I
want.

One slight problem that I need to use `:UpdateRemotePlugins`, but that's
something I can deal with.

The only real downside is that I have to press <TAB> twice in order to
get something up when the default is the value that I would have wanted.
This is fine.

Honestly not sure whether I'll use telescope any more, since everything
I wanted to do was something I know how to do on the vim command line.

Will try it and find out.