Remove the <C-b> insert mode mapping for <C-p>
I don't use it -- I thought that having the keybinding close to each
other on my Dvorak keyboard would be really useful, but I honestly
forgot it existed.

Now it's getting in the way of a few plugins, I strongly suspect the
problem would be solved by using `imap` instead of `inoremap`, but since
I don't use the mapping let's just remove it.