Add mappings to telescope for argument list handling
This helps managing the argument list in the nice (and fast) telescope
UI.  Mappings are to select/unselect everything, and to either set
arglist to everything selected or to add everything selected to the
argument list.

Trying to reduce my reliance on vimfindsme and vimple since they have a
few sore-spots and telescope seems like it will have more momentum than
those plugins.