Remove the :ls at the start of the <leader>b mapping
The idea behind this mapping was to print all buffers and get ready to
choose one.  I often end up with so many buffers that this keybinding
becomes unwieldy (the printing of buffers taking a lot of screen space).

When I'm not quite sure of the name or buffer number I use
fuzzy-finding for selection, and I only really want to use this mapping
when I know exactly where I want to get to.

Hence removing the "print" part of this mapping and just have it as an
easier to type version of :b (I know -- already pretty short, but this
change does indeed seem to change my habits too).