Use default selection modes for treesitter textobjects
I was trying to use linewise selection because I think it's "cleaner".
Turns out that linewise selection takes extra lines because the
textobject logically finishes after the trailing whitespace of a
function/class.

This meant that doing operations on a function/class would include more
than I actually wanted.

I think this is a bug -- and I think linewise selection would be really
useful in order to run commands over the relevant range -- however I can
also see that the charwise selection makes sense, and it works.

Using charwise selection (i.e. the default) for the relative lack of
bugs that comes with following the default.