Ugly script added to update submodules
Difference between this and `git submodule foreach` is that I need to
record a mapping between which project uses `main` and which uses
`master` (sigh ... differences).

Could have approached it in another way, but copy/paste from command
line and a few regexp search-replaces was the first way that came to
mind.