foldexpr use `setlocal` rather than `lua vim.wo.foldexpr`
Make the change at all because I had made a mistake in the lua setting
that meant even though I was setting a window option I was setting it
globally.  This was giving huge slowdown when opening a `tags` file
because the treesitter foldexpr was taking a lot of time.

I find `setlocal` a little easier to understand, and this means the
foldexpr is reset when moving away from a C buffer -- hence avoiding the
slowdown problems I was hitting.