Add command to remove multiple buffers at once
Honestly not sure whether this is useful -- I am currently using
telescope to select a bunch of buffers at once and remove them all and I
like this interface more.

It just "feels right" to have such a function, so will go with it since
it's just my config and I can do unreasonable things for the
satisfaction of it.