Avoid problem with helpers toggle_cursorcrosshairs
Was seeing that this would every so often lead to messed up cursor line
and column because I'd moved window and the callback was still
happening.

Avoid this properly with neovim by continuing to toggle that particular
windows crosshairs, without neovim simply do not allow the user to move
windows while this sequence of toggling is happening.