Add `j` to formatoptions in vimrc
While using original vim I've noticed that there's a difference between
it and neovim in the default values of `formatoptions`.

Neovim defaults to `tcqj` while vim defaults to `tcq`.  I only noticed
this in the `gitcommit` filetype (I guess because other ftplugins set
`j`).

Adjust the default in vimrc in order to have things match better.