Include colorscheme from github NLKNguyen/papercolor-theme
Currently working in a bright environment.  So starting terminals with
`xterm -bg white -fg black` and using this colorscheme.

N.b. If I directly switch to `PaperColor` from my own `techras` setting
I get the dark version of PaperColor.  Looks like this is something to
do with the setting of `background` to `dark` and some automatic choices
that PaperColor performs.

To work around this I use the below:
`:set background light | colorscheme PaperColor`