Various GUI improvements for techras colorscheme
Started to use `neovide` since my home office goes from so bright I need
a light colorscheme to not that bad throughout the day.  Tricks with a
"light xterm" and switching colorschemes on a vim session inside tmux
work, but are more awkward than *just* changing colorschemes in a GUI.

After working inside a GUI for the first time in a while, I see
improvements for my colorscheme.  Hence this change.